+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 0.2.0
+
+- `PatreonClient.fetchApplicationToken` is now deprecated and instead `PatreonCreatorClient.fetchApplicationToken` should be used.
+
+### Features
+
+- `PatreonClient`: `name` can be null
+- `PatreonClient`: add `refreshOnFailed` option to retry an unauthorized request once
+- `PatreonStore`: add Fetch and KV stores
+- clients: add `PatreonUserClient` and `PatreonCreatorClient`
+
 ## 0.1.0
 
 - Initial commit